+++ packages/nft-checkout/src/types/operation.ts
@@ -35,7 +35,7 @@ export interface INFTCheckoutOperation extends OperationSubscriber {
   init(args: OperationCreateParams): Promise<void>
 
   supportedChains(): Promise<BridgeChain[]>
-  supportedTokens(): Promise<Token[]>
+  supportedTokens(chain: BridgeChain): Promise<Token[]>
 
   loadPaymentTokens(chain: BridgeChain): Promise<PaymentToken[]>
   estimatePrice(token: PaymentToken): Promise<EstimatedPrice>