+++ packages/react-nft-checkout/src/components/PaymentTokensList/PaymentTokensList.tsx
@@ -17,6 +17,7 @@ const PaymentTokensList = () => {
     selectedChain,
     loadPaymentTokens,
     selectedPaymentToken,
+    selectedSwapToken,
     setSelectedPaymentToken,
     provider,
   } = useDappContext()
@@ -24,13 +25,13 @@ const PaymentTokensList = () => {
   const [isLoading, setIsLoading] = useState(true)
   const [loadingErrorText, setLoadingErrorText] = useState('')
   const [tokens, setTokens] = useState<PaymentToken[]>([])
-
+  const isDisabled = (token?: PaymentToken) =>
+    token?.symbol === selectedSwapToken?.symbol
   useEffect(() => {
     const fetchPaymentTokens = async () => {
       try {
         setIsLoading(true)
         setLoadingErrorText('')
-
         const paymentTokens = selectedChain
           ? (await loadPaymentTokens?.(selectedChain)) ?? []
           : []
@@ -80,6 +81,7 @@ const PaymentTokensList = () => {
                         }
                         key={paymentToken.symbol}
                         onClick={() => setSelectedPaymentToken(paymentToken)}
+                        disabled={isDisabled(paymentToken)}
                       >
                         <ListItemAvatar>
                           <Avatar>