+++ packages/react-nft-checkout/src/components/PriceConversion/PriceConversion.tsx
@@ -10,7 +10,7 @@ type Props = {
 }
 
 const PriceConversion = ({ isLoading, estimatedPrice }: Props) => {
-  const { targetNft, checkoutOperation } = useDappContext()
+  const { targetNft } = useDappContext()
 
   return (
     <>
@@ -39,7 +39,7 @@ const PriceConversion = ({ isLoading, estimatedPrice }: Props) => {
               >
                 <Typography variant="subtitle1">Price conversion</Typography>
                 <Typography variant="subtitle1">
-                  {`${targetNft.price} ${checkoutOperation?.chainFrom?.token.symbol} = ${estimatedPrice.price} ${estimatedPrice.price.symbol}`}
+                  {`${targetNft.price} ${estimatedPrice.to?.symbol} = ${estimatedPrice.price} ${estimatedPrice.price.symbol}`}
                 </Typography>
               </Box>
             </Box>