+++ packages/react-nft-checkout/src/components/SwapTokensSelect/SwapTokensSelect.tsx
@@ -1,0 +1,153 @@
+import {
+  FormControl,
+  InputLabel,
+  ListItemIcon,
+  ListItemText,
+  MenuItem,
+  Select,
+  SelectChangeEvent,
+} from '@mui/material'
+import { TARGET_TOKEN_SYMBOLS, Token } from '@rarimo/nft-checkout'
+import { EthProviderRpcError } from '@rarimo/provider'
+import { Network } from 'iconoir-react'
+import { useEffect, useMemo, useState } from 'react'
+
+import { ErrorText, LoadingIndicator } from '@/components'
+import styles from '@/components/BridgeChainSelect/BridgeChainSelect.module.css'
+import { useDappContext } from '@/hooks'
+
+const SwapTokensSelect = () => {
+  const {
+    getSupportedTokens,
+    provider,
+    selectedChain,
+    setSelectedSwapToken,
+    setSelectedPaymentToken,
+    selectedSwapToken,
+  } = useDappContext()
+
+  const isDisabled = (token?: Token) => {
+    if (token?.isNative) return true
+    for (const key in TARGET_TOKEN_SYMBOLS) {
+      if (TARGET_TOKEN_SYMBOLS[key] === token?.symbol) return true
+    }
+    return false
+  }
+  const handleChange = (event: SelectChangeEvent) => {
+    const selectedToken = tokens.find(
+      token => token.symbol === event.target.value,
+    )
+    setSelectedPaymentToken(null)
+    setSelectedSwapToken(selectedToken)
+  }
+
+  const value = useMemo(
+    () => (selectedSwapToken ? String(selectedSwapToken.symbol) : ''),
+    [selectedSwapToken],
+  )
+
+  const [isLoading, setIsLoading] = useState(true)
+  const [loadingErrorText, setLoadingErrorText] = useState('')
+  const [tokens, setTokens] = useState<Token[]>([])
+
+  useEffect(() => {
+    const fetchSupportedTokens = async () => {
+      try {
+        setIsLoading(true)
+        setLoadingErrorText('')
+        const supportedTokens = selectedChain
+          ? (await getSupportedTokens?.(selectedChain)) ?? []
+          : []
+        setTokens(supportedTokens)
+      } catch (error) {
+        setLoadingErrorText(
+          (error as unknown as EthProviderRpcError)?.message ||
+            'An error occurred while loading tokens. Change the network or try again later.',
+        )
+      }
+      setIsLoading(false)
+    }
+
+    fetchSupportedTokens()
+  }, [selectedChain, provider, getSupportedTokens])
+
+  useEffect(() => {
+    setSelectedSwapToken(undefined)
+
+    return () => {
+      setSelectedSwapToken(undefined)
+    }
+  }, [selectedChain, setSelectedSwapToken])
+
+  return (
+    <>
+      {loadingErrorText ? (
+        <ErrorText text={loadingErrorText} />
+      ) : (
+        <>
+          {isLoading ? (
+            <LoadingIndicator text="Loading supported tokens" />
+          ) : (
+            <>
+              {tokens.length ? (
+                <FormControl fullWidth sx={{ mt: 1 }}>
+                  <InputLabel id="token-select-lbl">Select a token</InputLabel>
+                  <Select
+                    fullWidth
+                    labelId="token-select-lbl"
+                    value={value}
+                    label="Select a token"
+                    onChange={handleChange}
+                  >
+                    {tokens.map(swapToken => {
+                      const tokenLogoUri = swapToken.logoURI?.replace(
+                        'ipfs://',
+                        'https://ipfs.io/ipfs/',
+                      )
+
+                      return (
+                        <MenuItem
+                          key={swapToken.symbol}
+                          value={swapToken.symbol}
+                          disabled={isDisabled(swapToken)}
+                        >
+                          <ListItemIcon
+                            sx={{
+                              minWidth: 40,
+                            }}
+                          >
+                            {swapToken.logoURI ? (
+                              <img
+                                className={styles.ListItemIcon}
+                                src={tokenLogoUri}
+                                width={28}
+                                height={28}
+                                alt={`${swapToken.name} icon`}
+                              />
+                            ) : (
+                              <Network className={styles.DefaultIcon} />
+                            )}
+                          </ListItemIcon>
+                          <ListItemText
+                            primary={swapToken.name}
+                            sx={{
+                              m: 0,
+                            }}
+                          />
+                        </MenuItem>
+                      )
+                    })}
+                  </Select>
+                </FormControl>
+              ) : (
+                <ErrorText text="No supported token." />
+              )}
+            </>
+          )}
+        </>
+      )}
+    </>
+  )
+}
+
+export default SwapTokensSelect