+++ CHANGELOG.md
@@ -10,10 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `@rarimo/react-nft-checkout`: React implementation of `@rarimo/nft-checkout` package
 - `@rarimo/react-nft-checkout`: Exported UI component `RarimoPayButon`
 - `@rarimo/react-nft-checkout`: Exported context provider `DappContextProvider` with core methods for NFT chekout
+- `@rarimo/nft-checkout`: Added selectable token to swap
+- `@rarimo/nft-checkout`: disabled native token and wrrapped native tokens
 
 ### Fixed
 - `@rarimo/nft-checkout`: Adding chain
 - `@rarimo/provider`: Adding chain
+- `@rarimo/provider`: load supported tokens
+
 
 ## [1.4.0] - 2023-03-13
 ### Added